Refactor : Continue applying Go's basic project structure
