+++ internal/types/UrlData.go
@@ -1,0 +1,8 @@
+package types
+
+/*
+* Data structure of the JSON coming from web client
+ */
+type URLData struct {
+	URL string `json:"url"`
+}
